+++ app/src/main/java/nextstep/payments/feature/cardlist/view/AddNewCardInfoText.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.view
+package nextstep.payments.feature.cardlist.view
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
